Extract bundle remove code
